delete -> delete & revert

+++ src/main/java/com/showka/web/u05/U05G002Controller.java
@@ -254,7 +254,7 @@ public class U05G002Controller extends ControllerBase {
 		// 掛売
 		Optional<Urikake> urikake = urikakeSpecificationService.buildUrikakeBy(uriage);
 		urikake.ifPresent(u -> {
-			// occ
+			// OCC
 			u.setVersion(form.getUrikakeVersion());
 			urikakeCrudService.save(u);
 		});
@@ -274,6 +274,7 @@ public class U05G002Controller extends ControllerBase {
 	 * delete.
 	 *
 	 */
+	// TODO revert delete 内部で分岐させるより初めからmethod分割しておくほうがよい。
 	@Transactional
 	@RequestMapping(value = "/u05g002/delete", method = RequestMethod.POST)
 	public ResponseEntity<?> delete(@ModelAttribute U05G002Form form, ModelAndViewExtended model) {
@@ -290,11 +291,18 @@ public class U05G002Controller extends ControllerBase {
 		// delete 商品移動
 		shohinIdoUriageCrudService.delete(pk);
 
-		// revert 売掛
-		urikakeCrudService.revert(form.getRecordId(), form.getUrikakeVersion());
-
-		// delete
-		uriageCrudService.delete(pk, form.getVersion());
+		// 売上履歴=1件 -> 一度も計上されていないで売上・売掛を削除。
+		// 売上履歴>1件 -> 前回計上状態に差し戻す。
+		UriageRireki rireki = uriageRirekiCrudService.getUriageRirekiList(form.getRecordId());
+		if (rireki.getList().size() == 1) {
+			// delete
+			urikakeCrudService.deleteIfExists(form.getRecordId(), form.getUrikakeVersion());
+			uriageCrudService.delete(pk, form.getVersion());
+		} else {
+			// revert
+			urikakeCrudService.revert(form.getRecordId(), form.getUrikakeVersion());
+			uriageCrudService.revert(pk, form.getVersion());
+		}
 
 		// message
 		form.setSuccessMessage("削除成功");
